requirements.txt
click==7.1.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
speedtest-cli==2.1.2
Werkzeug==1.0.1
